adds typing and refactors exports
